python_rag/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.5
botocore==1.35.5
certifi==2024.7.4
charset-normalizer==3.3.2
cohere==5.8.1
exceptiongroup==1.2.2
fastavro==1.9.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
jmespath==1.0.1
packaging==24.1
parameterized==0.9.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pydantic==2.8.2
pydantic-core==2.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-extensions==4.12.2
urllib3==2.2.2
